requirements.txt
fastapi
uvicorn[standard]
aiohttp
jinja2
python-dotenv
python-multipart
aiosqlite
